Rename `sparse_image_utils` to `sparse_image`
Bug: b/434763749
