+++ base/cvd/cuttlefish/host/commands/cvd/fetch/de_android_sparse.cc
@@ -25,7 +25,7 @@
 #include <sparse/sparse.h>
 
 #include "cuttlefish/common/libs/utils/result.h"
-#include "cuttlefish/host/libs/image_aggregator/sparse_image_utils.h"
+#include "cuttlefish/host/libs/image_aggregator/sparse_image.h"
 
 namespace cuttlefish {
 namespace {
