+++ base/cvd/cuttlefish/host/libs/image_aggregator/BUILD.bazel
@@ -38,7 +38,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/config:mbr",
         "//cuttlefish/host/libs/image_aggregator:cdisk_spec_cc_proto",
         "//cuttlefish/host/libs/image_aggregator:qcow2",
-        "//cuttlefish/host/libs/image_aggregator:sparse_image_utils",
+        "//cuttlefish/host/libs/image_aggregator:sparse_image",
         "//libbase",
         "//libsparse",
         "@protobuf",
@@ -47,15 +47,13 @@ cf_cc_library(
 )
 
 cf_cc_library(
-    name = "sparse_image_utils",
-    srcs = ["sparse_image_utils.cc"],
-    hdrs = ["sparse_image_utils.h"],
+    name = "sparse_image",
+    srcs = ["sparse_image.cc"],
+    hdrs = ["sparse_image.h"],
     deps = [
         "//cuttlefish/common/libs/fs",
         "//cuttlefish/common/libs/utils:result",
         "//cuttlefish/common/libs/utils:subprocess",
-        "//cuttlefish/host/libs/config:config_utils",
-        "//cuttlefish/host/libs/config:cuttlefish_config",
         "//cuttlefish/host/libs/config:known_paths",
         "//libbase",
     ],
