+++ base/cvd/cuttlefish/host/libs/image_aggregator/sparse_image.h
